+++ django_cleanup/apps.py
@@ -9,7 +9,16 @@ from . import cache, handlers
 class CleanupConfig(AppConfig):
     name = 'django_cleanup'
     verbose_name = 'Django Cleanup'
+    default = True
 
     def ready(self):
         cache.prepare()
+        handlers.connect()
+
+class CleanupSelectedConfig(AppConfig):
+    name = 'django_cleanup'
+    verbose_name = 'Django Cleanup'
+
+    def ready(self):
+        cache.prepare(True)
         handlers.connect()